Fix compile warning on Solaris, per buildfarm.  (Why have we got three slightly different copies of this file?)

+++ contrib/hstore/crc32.c
@@ -90,7 +90,7 @@ static const unsigned int crc32tab[256] = {
 unsigned int
 crc32_sz(char *buf, int size)
 {
-	unsigned int crc = ~0;
+	unsigned int crc = ~((unsigned int) 0);
 	char	   *p;
 	int			len,
 				nr;
