+++ contrib/ltree/crc32.c
@@ -1,6 +1,6 @@
 /* Both POSIX and CRC32 checksums */
 
-/* $PostgreSQL: pgsql/contrib/ltree/crc32.c,v 1.7 2006/09/22 21:39:57 tgl Exp $ */
+/* $PostgreSQL: pgsql/contrib/ltree/crc32.c,v 1.8 2007/07/15 22:40:28 tgl Exp $ */
 
 #include <sys/types.h>
 #include <stdio.h>
@@ -99,7 +99,7 @@ static const unsigned int crc32tab[256] = {
 unsigned int
 ltree_crc32_sz(char *buf, int size)
 {
-	unsigned int crc = ~0;
+	unsigned int crc = ~((unsigned int) 0);
 	char	   *p;
 	int			len,
 				nr;
